requirements.txt
streamlit>=1.23.0
fpdf>=1.7.2
matplotlib>=3.4.3
numpy>=1.19.5
openai==0.28.0
requests>=2.25.1
Pillow>=8.2.0
python-dotenv==1.0.0
uuid==1.30
